Add SConstruct and SConscript for building with SCons (work in progress)
The motivation for this is that I work on a Mac and want to get rid of
my Windows VM with Atmel Studio. avr-gcc can be used on a Mac (I chose
CrossPack for AVR because it also includes avr-gdb.

Currently building with SCons fails due to a missing symbol for the heap
implementation.